requirements.txt
selenium>=4.20.0
pandas>=2.2.0
openpyxl>=3.1.2
colorama>=0.4.6
customtkinter>=5.2.2
python-dotenv>=1.0.1